// client/src/app/components/enum.ts
// interface enum
export enum TOOLS {
  PENCIL = 'pencil',
  DRAW = 'draw',
  LINE = 'line',
  RECT = 'rectangle',
  COLOR = 'color',
  TEXT = 'text',
  ERASER = 'eraser',
  SHAPE = 'shape',
  BRUSH = 'brush',
  SELECTOR = 'selection',
  APPLICATOR = 'applicator',
  ELLIPSE = 'ellipse',
  POLYGON = 'polygon',
  BUCKET = 'bucket',
  DROPPER = 'dropper',
  SPRAY = 'spray',
  GRID = 'grid',
  FEATHER = 'feather',
}

export enum KEY {
  I = 'i',
  C = 'c',
  L = 'l',
  W = 'w',
  STR1 = '1',
  STR2 = '2',
  STR3 = '3',
  SHIFT = 'Shift',
  BACKSPACE = 'Backspace',
  ESCAPE = 'Escape',
  DELETE = 'Delete',
  ENTER = 'Enter',
  S = 's',
  RIGHT_ARROW = 'ArrowRight',
  LEFT_ARROW = 'ArrowLeft',
  UP_ARROW = 'ArrowUp',
  DOWN_ARROW = 'ArrowDown',
  E = 'e',
  R = 'r',
  G = 'g',
  A = 'a',
  T = 't',
  B = 'b',
  P = 'p',
}

export enum TEXTURE {
  TURBULENCE = 'turbulence',
  BLURRY = 'blurry',
  BASE = 'base',
  NOISE = 'noise',
  SQUIGLY = 'squigly',
}

export enum DRAW {
  PENCIL = 'pencil',
  BASESTROKE = 5,
  BASESTROKESPRAY = 20
}

export enum COLOR {
  BUCKET = 'bucket'
}

export enum SHAPE {
  BASESTROKEWIDTH = 10,
}

export enum SUBJECT {
  NONE = 'none',
  WELCOME = 'welcome',
  PENCIL = 'Crayon',
  BRUSH = 'Pinceau',
  LINE = 'Ligne',
  AIRBRUSH = 'Aérosol',
  RECT = 'Rectangle',
  ELLIPSE = 'Ellipse',
  POLYGON = 'Polygone',
  BUCKET = 'Sceau de peinture',
  PIPETTE = 'Pipette',
  APPLICATOR = 'Applicateur de couleur',
  SELECTION = 'Sélection',
  GRID = 'Grille',
  CREATEDRAWING = 'Créer un dessin',
  EXPORTDRAWING = 'Exporter un dessin',
  GALLERY = 'Galerie',
  SAVE = 'Sauvegarder',
  ERASOR = 'Efface',
  TEXT = 'Écrire un texte',
}

export enum BUTTONCATEGORY {
  TOOLS = 'tools',
  OPTIONS = 'options',
}